requirements.txt
torch==2.7.1+cu128
torchvision==0.22.1+cu128
matplotlib==3.10.3
numpy==2.2.6
pillow==11.2.1
tqdm==4.67.1